requirements.txt
easyocr
streamlit
pandas
openpyxl
Pillow